fix: nav popups
